Separate into getTransferFeedParams
